Managed to get a basic working version by following Jawgs tutorial. They were using script tags so just did taht for a bit to see waht the map will look like. Looks v cool. Need to get to grips with Express in order to make server side JS that can call Jawgs API
