+++ css/styles.css
@@ -1,5 +1,6 @@
 #map {
-    height: 180px;
+    width: 600px;
+    height: 360px;
     border: 5px dashed rgb(250, 5, 136);
 }
