+++ index.html
@@ -25,7 +25,7 @@
 
     </div> 
     <!-- end of giant container -->
-
+    
     <!-- Leaflet JS -->
     <script src="https://unpkg.com/leaflet@1.5.1/dist/leaflet.js"
     integrity="sha512-GffPMF3RvMeYyc1LWMHtK8EbPv0iNZ8/oTtHPx9/cc2ILxQ+u905qIwdpULaqDkyBKgOaB57QTMg7ztg8Jm2Og=="
@@ -42,5 +42,6 @@
     crossorigin="anonymous"></script>
     <!-- my JS -->
     <script src="./js/frontend.js"></script>
+    
 </body>
 </html>
